+++ linuxdoom-1.10/r_data.cxx
@@ -24,7 +24,7 @@
 //-----------------------------------------------------------------------------
 
 
-#include <malloc.h>
+#include <stdlib.h>
 static const char
 rcsid[] = "$Id: r_data.c,v 1.4 1997/02/03 16:47:55 b1 Exp $";
 
@@ -630,7 +630,7 @@ void R_InitColormaps (void)
     lump = W_GetNumForName("COLORMAP"); 
     length = W_LumpLength (lump) + 255; 
     colormaps = (lighttable_t*)malloc(length); 
-    colormaps = (byte *)( ((int)colormaps + 255)&~0xff); 
+    colormaps = (byte *)( ((intptr_t)colormaps + 255)&~0xff); 
     W_ReadLump (lump,colormaps); 
 }
